requirements.txt
anyio==3.7.1
appdirs==1.4.4
Babel==2.12.1
certifi==2023.5.7
Cython==0.29.21
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
joblib==1.2.0
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
PyYAML==5.3.1
rfc3986==1.5.0
ruamel.yaml==0.17.4
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
threadpoolctl==2.1.0
typing_extensions==4.7.1
tzdata==2023.3
